+++ src/components/PlayerView/playerViewConfig.ts
@@ -27,6 +27,18 @@ export interface PlayerViewConfig {
    * To reliably hide the first frame before a pre-roll ad, please ensure that you are using the {@link AdvertisingConfig} to schedule ads and not the {@link Player.scheduleAd} API call.
    */
   hideFirstFrame?: boolean;
+
+  /**
+   * Specify on which surface type the video should be rendered.
+   *
+   * See {@link https://developer.android.com/guide/topics/media/ui/playerview#surfacetype|Choosing a surface type}
+   * for more information.
+   *
+   * Default is {@link SurfaceType.SurfaceView}.
+   *
+   * @platform Android
+   */
+  surfaceType?: SurfaceType;
 }
 
 /**
@@ -96,4 +108,20 @@ export class CustomUi extends Variant {
   constructor(uiManagerFactoryFunction: string) {
     super(uiManagerFactoryFunction);
   }
+}
+
+/**
+ * The type of surface on which to render video.
+ *
+ * See {@link https://developer.android.com/guide/topics/media/ui/playerview#surfacetype|Choosing a surface type}
+ * for more information.
+ */
+export enum SurfaceType {
+  /**
+   * SurfaceView generally causes lower battery consumption,
+   * and has better handling for HDR and secure content.
+   */
+  SurfaceView,
+  /** TextureView is sometime needed for smooth animations. */
+  TextureView,
 }